pyproject.toml
[project]
name = "mana-aatalu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "streamlit>=1.48.1",
]
